With 4.0.0 release, blur filter algorithm have been changed, and radius argument precision have been dropped due to use an interger value intead a float value. This commit fix this wrong cast to take a acre about radius precision with UnsharpMask filter BUGS: 335666 FIXED-IN: 4.1.0

+++ libs/dimg/filters/sharp/unsharpmaskfilter.h
@@ -41,7 +41,7 @@ class DIGIKAM_EXPORT UnsharpMaskFilter : public DImgThreadedFilter
 public:
 
     explicit UnsharpMaskFilter(QObject* const parent = 0);
-    explicit UnsharpMaskFilter(DImg* const orgImage, QObject* const parent=0, int radius=1,
+    explicit UnsharpMaskFilter(DImg* const orgImage, QObject* const parent=0, double radius=1.0,
                                double amount=1.0, double threshold=0.05);
 
     // Constructor for slave mode: execute immediately in current thread with specified master filter
@@ -85,7 +85,7 @@ private:
 
 private:
 
-    int    m_radius;
+    double m_radius;
 
     double m_amount;
     double m_threshold;
